feat: マイグレーションファイルをマイグレート #36

+++ db/migrate/20230812132430_create_projects.rb
@@ -4,7 +4,7 @@ class CreateProjects < ActiveRecord::Migration[6.1]
       t.string :name, null: false
       t.date :start_date, null: false
       t.date :end_date, null: false
-      t.smallint :status, null: false, default: 0
+      t.integer :status, null: false, default: 0
       t.string :overview
 
       t.timestamps
